// src/data/airports.generated.ts
export type AirportOption = {
  iata: string;
  name: string;
  city: string;
  search: string;
};

export type AirportCountryGroup = {
  code: string;
  country: string;
  countryEnglish: string;
  region: "europe" | "world";
  airports: AirportOption[];
};